pyproject.toml
[project]
name = "ai_lab"
version = "0.2.0"
description = ""
authors = [
    {name = "Dmitriy Lavrinovich",email = "dlavrinovich@hse.ru"}
]
requires-python = ">=3.13,<3.14"
dependencies = [
    "spacy (>=3.8.7,<4.0.0)",
    "sentence-transformers (>=5.0.0,<6.0.0)",
    "bertopic (>=0.17.3,<0.18.0)",
    "transformers (>=4.53.1,<5.0.0)",
    "torch (>=2.7.1,<3.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "umap-learn (>=0.5.9.post2,<0.6.0)",
    "scikit-learn (>=1.7.0,<2.0.0)",
    "streamlit (>=1.46.1,<2.0.0)",
    "plotly (>=5.21.0,<6.0.0)",
    "colorama (>=0.4.6,<1.0.0)",
    "altair (>=5.5.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.2"

[tool.ruff]
preview = true
